requirements.txt
# Core RAG dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6

# Document processing
PyPDF2>=3.0.1
beautifulsoup4>=4.12.2
markdown>=3.5.1
python-docx>=1.1.0
Pillow>=10.1.0

# Vector embeddings & search
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.3
scikit-learn>=1.3.2

# Database
sqlalchemy>=2.0.23

# LLM integration
openai>=1.3.7
anthropic>=0.7.8
google-generativeai>=0.3.2

# Text processing
python-dateutil>=2.8.2

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
aiofiles>=23.2.1
tqdm>=4.66.1

# Production dependencies
gunicorn>=21.2.0
